Refactor : OrgParty and AttParty now extend Party
