+++ cpmac/app/shared/lib/dedjs/object/pop/att/AttParty.js
@@ -1,6 +1,5 @@
 const Helper = require("../../../Helper");
 const ObservableModule = require("data/observable");
-const ObservableArray = require("data/observable-array").ObservableArray;
 const KeyPair = require("../../../Crypto").KeyPair;
 const FileIO = require("../../../../file-io/file-io");
 const FilePath = require("../../../../../res/files/files-path");
@@ -9,16 +8,14 @@ const Net = require("@dedis/cothority").net;
 const CothorityMessages = require("../../../network/cothority-messages");
 const RequestPath = require("../../../network/RequestPath");
 const DecodeType = require("../../../network/DecodeType");
+const Party = require("../Party");
 
-/**
- * This singleton represents the attendee of a PoP party. It contains everything related to the attendee.
- */
 
 /**
  * We define the AttParty class which is the object representing the attendee.
  */
 
-class AttParty {
+class AttParty extends Party {
 
   /**
    * Constructor for the AttParty class.
@@ -29,6 +26,7 @@ class AttParty {
    *
    */
   constructor(id, loadLocally, address) {
+    super();
     if (typeof id !== "string" || id === "") {
       throw new Error("id must be of type string and shouldn't be empty");
     }
@@ -45,16 +43,6 @@ class AttParty {
     this._isLoaded = false;
     this._finalStatement = undefined;
     this._keyPair = new KeyPair(FileIO.join(FilePath.POP_ATT_PATH, this._folderName));
-    this._popDesc = ObservableModule.fromObjectRecursive({
-      name: "",
-      dateTime: "",
-      location: "",
-      roster: {
-        id: new Uint8Array(),
-        list: new ObservableArray(),
-        aggregate: new Uint8Array()
-      }
-    });
     this._status = ObservableModule.fromObject({
       status: States.UNDEFINED
     });
@@ -205,6 +193,12 @@ class AttParty {
     }
   }
 
+  /**
+   * Check if a specific public key is part of the party
+   *
+   * @param publicKey - the public that has to be checked
+   * @return {boolean} - returns true if the publix key is in the party
+   */
   isAttendee(publicKey) {
     let attendees = this._finalStatement.attendees;
     let publicKeyHexString = Convert.byteArrayToHex(publicKey);
@@ -215,14 +209,6 @@ class AttParty {
     }
 
     return false;
-  }
-
-  /**
-   * Returns the observable module for the description of the party
-   * @returns {ObservableModule}
-   */
-  getPopDescModule() {
-    return this._popDesc;
   }
 
   /**
